Implementing consumption models for workstations.

+++ src/main/java/be/kuleuven/cs/flexsim/domain/workstation/CurtailableStationDecorator.java
@@ -62,7 +62,7 @@ public class CurtailableStationDecorator extends ForwardingStationDecorator
     }
 
     @Override
-    public int getLastStepConsumption() {
+    public double getLastStepConsumption() {
         if (isCurtailed()) {
             return 0;
         }
